\name{biasVardistPsi0Compute}
\Rdversion{1.0}
\alias{biasVardistPsi0Compute}
\title{one line description}
\description{
  description
}
\usage{
Psi0 <- biasVardistPsi0Compute(biasKern, vardist)
}
\arguments{
  \item{biasKern}{the kernel structure associated with the white kernel.}
  \item{vardist}{description}
}
\value{
  \item{Psi0}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
