\name{pca}
\Rdversion{1.0}
\alias{pca}
\title{Principal Components Analysis}
\description{
   PCCOEFF = PCA(DATA) computes the eigenvalues of the covariance matrix of the dataset DATA and returns them as PCCOEFF.  These coefficients give the variance of DATA along the corresponding principal components.
}
\usage{
pca()
}
\arguments{
}
\value{
}
\keyword{model}
