\name{priorParamInit}
\Rdversion{1.0}
\alias{priorParamInit}
\title{Prior model's parameter initialisation.}
\description{
  wrapper function for initialising prior distributions parameters.
}
\usage{
prior <- priorParamInit(prior)
}
\arguments{
  \item{prior}{structure to initialise.}
}
\value{
  \item{prior}{initialised prior structure.}
}
\seealso{
\code{\link{priorCreate}}.
}
\keyword{model}
