\name{scg}
\Rdversion{1.0}
\alias{scg}
\title{Scaled conjugate gradient optimization.}
\description{
   [X, OPTIONS] = SCG(F, X, OPTIONS, GRADF) uses a scaled conjugate gradients algorithm to find a local minimum of the function F(X) whose gradient is given by GRADF(X).  Here X is a row vector and F returns a scalar value. The point at which F has a local minimum is returned as X.  The function value at that point is returned in OPTIONS(8).
}
\usage{
scg()
}
\arguments{
}
\value{
}
\keyword{model}
