\name{whiteVardistPsi0Compute}
\Rdversion{1.0}
\alias{whiteVardistPsi0Compute}
\title{one line description}
\description{
  description
}
\usage{
Psi0 <- whiteVardistPsi0Compute(whiteKern, vardist)
}
\arguments{
  \item{whiteKern}{the kernel structure associated with the white kernel.}
  \item{vardist}{description}
}
\value{
  \item{Psi0}{description}
}
\seealso{
\code{\link{others}}.
}
\keyword{model}
